@@ -1,18 +1,21 @@
-import React, { Component, Fragment } from "react";
-import { Route, withRouter, Switch } from "react-router-dom";
+import React, { Fragment, useCallback, useEffect, useState } from "react";
+import { Route, Routes } from "react-router-dom";
 import { connect } from "react-redux";
-import { CartService } from "./services";
-import Meeting from './pages/home/components/videoCall/Meeting';
+// import { CartService } from "./services";
+// import Meeting from './pages/home/components/videoCall/Meeting';
 
-import { Header, Footer, Appbar, HeaderMessage } from "./shared";
+import Header from "./components/header/header";
+import HeaderMessage from "./components/header/headerMessage";
+import Appbar from "./components/header/appbar";
+import Footer from "./components/footer/footer";
 import {
   Home,
   List,
-  MyCoupons,
+  // MyCoupons,
   Detail,
   SuggestedProductsList,
   Profile,
-  ShoppingCart,
+  // ShoppingCart,
   Arrivals,
   RefundPolicy,
   TermsOfService,
@@ -21,118 +24,105 @@ import {
   Cart,
 } from "./pages";
 
-import "./i18n";
+// import "./i18n";
 import "./main.scss";
 import warningIcon from './assets/images/original/Contoso_Assets/Icons/information_icon.svg'
+import { useLocation } from "react-router-dom";
+import { CartService } from "./services";
+import { getCartQuantity } from "./actions/actions";
 
-class App extends Component {
-  constructor() {
-    super();
-    this.state = {
-      shoppingCart: [],
-      quantity: null,
-    };
-  }
 
-  async componentDidMount() {
-    if (this.props.userInfo.token) {
-      const shoppingCart = await CartService.getShoppingCart(
-        this.props.userInfo.token
-      );
-      if (shoppingCart) {
-        this.setState({ shoppingCart });
+  function App(props) {
+    const location = useLocation()
+    // const [shoppingCart, setShoppingCart] = useState([])
+    const [quantity, setQuantity] = useState(0)
+
+    const getQuantity = useCallback(async() => {
+      let quantity = 0;
+      //Show cart using API
+      if (props.userInfo.token) {
+        const shoppingcart = await CartService.getShoppingCart(
+          props.userInfo.token
+        );
+        // if (shoppingcart) {
+        //   setShoppingCart({ shoppingcart });
+        // }
+        quantity = shoppingcart.length;
+      }else{
+        let cart = localStorage.getItem('cart_items') ? JSON.parse(localStorage.getItem('cart_items')) : [];
+        quantity = cart.length;
       }
-    }
+      setQuantity(quantity);
+    },[props])
+    
+    useEffect(() => {
+      props.getCartQuantity(quantity)
+    }, [quantity, props]);
 
-    if (this.state.shoppingCart != null) {
-      const quantity = this.state.shoppingCart.reduce(
-        (oldQty, { qty }) => oldQty + qty,
-        0
-      );
-      this.setState({ quantity });
-    }
-  }
+    React.useEffect(() => {
+      getQuantity()
+    }, [getQuantity]);
 
-  componentDidUpdate(prevProps, prevState) {
-    if(this.props.location.pathname !== prevProps.location.pathname){
-      window.scrollTo(0, 0);
-    }
-  }
-
-  ShoppingCart = (quantity) => {
-    this.setState({ quantity });
-  };
-
-  sumProductInState = () => {
-    this.setState((prevState) => {
-      return { quantity: prevState.quantity + 1 };
-    });
-  };
-
-  render() {
-    const { quantity } = this.state;
-
-    const PrivateRoute = ({ component: Component, ...rest }) => (
-      <Route
-        {...rest}
-        render={(props) =>
-          this.props.userInfo.loggedIn === true ? (
-            <Component {...props} {...rest} />
-          ) : (
-            this.props.history.push('/')
-          )
-        }
-      />
-    );
+    React.useEffect(() => {
+        window.scrollTo(0, 0);
+    }, [location.pathname]);
 
     return (
-      <div className="App">
+      <div className={`App ${props.theme ? 'dark' : 'light' }`}>
         <Fragment>
           <div className="mainHeader">
             <HeaderMessage type="warning" icon={warningIcon} message="This Is A Demo Store For Testing Purposes — No Orders Shall Be Fulfilled."/>
-            <Appbar quantity={quantity} />
-            {this.props.history.location.pathname === '/' || this.props.history.location.pathname === '/new-arrivals' ?
-              <Header quantity={quantity} />
+            <Appbar />
+            {location.pathname === '/' || location.pathname === '/new-arrivals' ?
+              <Header/>
               :
               <div id="box"></div>}
           </div>
-          <Switch>
-            <Route exact path="/" component={Home} />
-            <Route exact path="/new-arrivals" component={Arrivals} />
-            <Route exact path="/meeting" component={Meeting} />
-            <Route exact path="/list" component={List} />
-            <Route exact path="/list/:code" component={List} />
+          <Routes>
+            <Route exact path="/" element={<Home/>} />
+            <Route exact path="/new-arrivals" element={<Arrivals/>} />
+            {/* <Route exact path="/meeting" element={Meeting} /> */}
+            <Route exact path="/list" element={<List/>} />
+            <Route exact path="/list/:code" element={<List/>} />
             <Route
               path="/suggested-products-list"
-              component={SuggestedProductsList}
+              element={<SuggestedProductsList/>}
             />
             <Route
               path="/product/detail/:productId"
-              render={(props) => (
-                <Detail sumProductInState={this.sumProductInState} {...props} />
-              )}
+              element={<Detail/>}
             />
-            <Route path="/refund-policy" component={RefundPolicy} />
-            <Route path="/terms-of-service" component={TermsOfService} />
-            <Route path="/about-us" component={AboutUs} />
-            <PrivateRoute path="/coupons" component={MyCoupons} />
-            <PrivateRoute path="/profile/:page" component={Profile} />
-            <PrivateRoute path="/cart" component={Cart}/>
-            <PrivateRoute
+            <Route path="/refund-policy" element={<RefundPolicy/>} />
+            <Route path="/terms-of-service" element={<TermsOfService/>} />
+            <Route path="/about-us" element={<AboutUs/>} />
+            {/* <PrivateRoute path="/coupons" element={MyCoupons} /> */}
+            {props.userInfo.loggedIn === true ?
+            <>
+            <Route path="/profile/:page" element={<Profile/>} />
+            </>:null}
+            <Route path="/cart" element={<Cart/>}/>
+            {/* <PrivateRoute
               path="/shopping-cart"
-              component={ShoppingCart}
+              element={ShoppingCart}
               ShoppingCart={this.ShoppingCart}
               quantity={this.state.quantity}
-            />
-            <Route path="*" component={ErrorPage} />
-          </Switch>
+            /> */}
+            <Route path="*" element={<ErrorPage/>} />
+          </Routes>
           <Footer />
         </Fragment>
       </div>
     );
   }
-}
+// }
 
-const mapStateToProps = (state) => state.login;
-
-export default withRouter(connect(mapStateToProps)(App));
+const mapStateToProps = (state) => { 
+  return { 
+    userInfo : state.login.userInfo,
+    theme :  state.login.theme
+  }
+};
+const mapDispatchToProps = (dispatch) => ({
+  getCartQuantity: (value) => dispatch(getCartQuantity(value)),
+})
+export default connect(mapStateToProps, mapDispatchToProps)(App);
