@@ -1,12 +1,13 @@
-import { Button } from '@material-ui/core';
+import { Button } from '@mui/material';
 import React from 'react';
 import errorpic from "../../assets/images/original/Contoso_Assets/404_page_assets/404_image.svg";
-import { useHistory } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
+import './errorPage.scss'
 
 const ErrorPage = () => {
-    const history = useHistory();
+    const history = useNavigate();
     const redirectTo = (path) => {
-        history.push(path)
+        history(path)
     }
     return (
         <>
