@@ -1,12 +1,12 @@
-import { Grid } from "@material-ui/core";
+import { Grid } from "@mui/material";
 import React from "react";
-import { OfferBanner, ListGrid, ListAside } from "./components";
-import Breadcrump  from "../../components/breadcrumb";
-import { withRouter, useHistory } from "react-router-dom";
+import { OfferBanner, ListGrid, ListAside } from "./sections";
+import Breadcrump  from "../../components/breadcrumb/breadcrumb";
+import { useLocation } from "react-router-dom";
 
 const List = ({ typesList, brandsList, onFilterChecked, productsList, loggedIn }) => {
-    const history = useHistory();
-    const currentCategory = history.location.pathname.split("/").pop().replaceAll('-',' ');
+    const location = useLocation();
+    const currentCategory = location.pathname.split("/").pop().replaceAll('-',' ');
 
     return (
         <div className="list">
@@ -18,14 +18,14 @@ const List = ({ typesList, brandsList, onFilterChecked, productsList, loggedIn }
             <div className="list__content">
                 <h6 className="mainHeading">{currentCategory}</h6>
                 <Grid container>
-                    <Grid item xs={3}>
+                    <Grid item lg={3} xs={12}>
                         <ListAside
                             onFilterChecked={onFilterChecked}
                             typesList={typesList}
                             brandsList={brandsList}
                         />
                     </Grid>
-                    <Grid item xs={9}>
+                    <Grid item lg={9} xs={12}>
                         <ListGrid productsList={productsList} />
                     </Grid>
                 </Grid>
@@ -35,4 +35,4 @@ const List = ({ typesList, brandsList, onFilterChecked, productsList, loggedIn }
     );
 };
 
-export default withRouter(List);
+export default (List);
