@@ -1,20 +1,24 @@
-import React from "react";
-import ReactDOM from "react-dom";
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import './index.css';
+import App from './App';
+import reportWebVitals from './reportWebVitals';
 import { Provider } from 'react-redux';
-import "./index.css";
-import App from "./App";
-import './i18n';
 import store from './store';
-import ConfigService from './services/configService';
-import { Router } from "react-router-dom";
-import { createBrowserHistory } from "history";
+import { BrowserRouter } from "react-router-dom";
 
-const history = createBrowserHistory({ basename: "" });
-(async () => {
-    await ConfigService.loadSettings();
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+  <React.StrictMode>
+    <Provider store={store}>
+      <BrowserRouter>
+        <App />
+      </BrowserRouter>
+    </Provider>
+  </React.StrictMode>
+);
 
-    ReactDOM.render(
-        <Provider store={store}><Router history={history}><App /></Router></Provider>,
-        document.getElementById('root')
-    );
-})();
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals();
