@@ -1,9 +1,9 @@
 import React from 'react';
-import Breadcrump from '../../components/breadcrumb.js'
-import { useHistory } from 'react-router-dom';
+import Breadcrump from '../../components/breadcrumb/breadcrumb'
+import { useLocation } from 'react-router-dom';
 const TermsOfService = (props) => {
-    const history = useHistory();
-    const currentCategory = history.location.pathname.split("/").pop().replaceAll('-',' ');
+    const location = useLocation();
+    const currentCategory = location.pathname.split("/").pop().replaceAll('-',' ');
     return (
         <>
             <div className='refund-policy-section'>
@@ -15,7 +15,7 @@ const TermsOfService = (props) => {
                         <b>Introduction</b>
                         <br/>
                         These Website Standard Terms and Conditions written on this webpage shall manage your use of our website, Website Name accessible at Contoso traders. 
-                        <br/>
+                        <br/><br/>
                         By using our website, you accepted these terms and conditions in full. If you disagree with these terms and conditions or any part of these terms and conditions, you must not use our website. 
                         <br/><br/>
                         <b>Intellectual Property Rights </b>
@@ -46,15 +46,15 @@ const TermsOfService = (props) => {
                         <br/><br/>
                         <b>Acceptable use </b>
                         <br/>
-                        You must not use our website in any way that causes, or may cause, damage to the website or impairment of the availability or accessibility of the website; or in any way which is unlawful, illegal, fraudulent, or harmful, or in connection with any unlawful, illegal, fraudulent, or harmful purpose or activity. 
-                        <br/>
-                        You must not use our website to copy, store, host, transmit, send, use, publish or distribute any material which consists of (or is linked to) any spyware, computer virus, Trojan horse, worm, keystroke logger, rootkit, or other malicious computer software. 
-                        <br/>
-                        You must not conduct any systematic or automated data collection activities (including without limitation scraping, data mining, data extraction and data harvesting) on or in relation to our website without our express written consent. 
-                        <br/>
-                        You must not use our website to transmit or send unsolicited commercial communications. 
-                        <br/>
-                        You must not use our website for any purposes related to marketing without our express written consent. 
+                        - You must not use our website in any way that causes, or may cause, damage to the website or impairment of the availability or accessibility of the website; or in any way which is unlawful, illegal, fraudulent, or harmful, or in connection with any unlawful, illegal, fraudulent, or harmful purpose or activity. 
+                        <br/><br/>
+                        - You must not use our website to copy, store, host, transmit, send, use, publish or distribute any material which consists of (or is linked to) any spyware, computer virus, Trojan horse, worm, keystroke logger, rootkit, or other malicious computer software. 
+                        <br/><br/>
+                        - You must not conduct any systematic or automated data collection activities (including without limitation scraping, data mining, data extraction and data harvesting) on or in relation to our website without our express written consent. 
+                        <br/><br/>
+                        - You must not use our website to transmit or send unsolicited commercial communications. 
+                        <br/><br/>
+                        - You must not use our website for any purposes related to marketing without our express written consent. 
                         <br/><br/>
 
                         <b>Restricted access </b>
@@ -64,9 +64,9 @@ const TermsOfService = (props) => {
                         <b>User content </b>
                         <br/>
                         In these terms and conditions, “your user content” means material (including without limitation text, images, audio material, video material and audio-visual material) that you submit to our website, for whatever purpose. 
-                        <br/>
+                        <br/><br/>
                         You grant to us a worldwide, irrevocable, non-exclusive, royalty-free license to use, reproduce, adapt, publish, translate, and distribute your user content in any existing or future media. You also grant to us the right to sub-license these rights, and the right to bring an action for infringement of these rights. 
-                        <br/>
+                        <br/><br/>
                         Your user content must not be illegal or unlawful, must not infringe any third party's legal rights, and must not be capable of giving rise to legal action whether against you or us or a third party (in each case under any applicable law). 
                         <br/>
                     <br/></p>
