@@ -1,95 +1,112 @@
-import React, { Component, Fragment } from 'react';
-import { withRouter } from 'react-router-dom';
-import { LoadingSpinner } from '../../shared/index';
-
+import React, { Fragment } from 'react';
+// import { withRouter } from 'react-router-dom';
+import LoadingSpinner from '../../components/loadingSpinner/loadingSpinner';
+import './list.scss'
 import List from './list';
 import { ProductService } from '../../services';
+import { useParams } from 'react-router-dom';
+// class ListContainer extends Component {
+  // constructor(props) {
+  //   super(props);
+  //   this.state = {
+  //     typesList: [],
+  //     brandsList: [],
+  //     productsList: [],
+  //     queryString: '',
+  //     loading: true,
+  //   };
 
-class ListContainer extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      typesList: [],
-      brandsList: [],
-      productsList: [],
-      queryString: '',
-      loading: true,
-    };
+  //   this.queryString = [
+  //     {
+  //       brand: [],
+  //       type: [],
+  //     },
+  //   ];
+  //   this.type = [];
+  //   this.brand = [];
+  // }
 
-    this.queryString = [
+  
+  const brand = []
+  function ListContainer() {
+    const [typesList, setTypesList] = React.useState([]);
+    const [brandsList, setBrandsList] = React.useState([]);
+    const [productsList, setProductsList] = React.useState('');
+    const queryString = 
       {
-        brand: [],
-        type: [],
-      },
-    ];
-    this.type = [];
-    this.brand = [];
-  }
+        brand: brand,
+        type: '',
+      }
+    ;
+    const [loading, setLoading] = React.useState(true);
+    const [getType, setType] = React.useState([]);
+    // const type = []
+    const { code } = useParams(); 
+    // React.useEffect(() => {
+    //   const filter = code || '';
+    //   getProductData(code);
+    //   // setPageState(filteredProductsPageData);
+    // }, []);
+  
+    React.useEffect(() => {
+        getProductData(code);
+    }, [code]);// eslint-disable-line react-hooks/exhaustive-deps
 
-  async componentDidMount() {
-    const filter = this.props.match.params.code || '';
-    const filteredProductsPageData = await this.getProductData(filter);
-    this.setPageState(filteredProductsPageData);
-  }
-
-  async componentDidUpdate(prevProps) {
-    const code = this.props.match.params.code;
-    if (code !== prevProps.match.params.code) {
-      const filteredProductsPageData = await this.getProductData(code);
-      this.setPageState(filteredProductsPageData);
+  
+    const getProductData = async(type) => {
+      setType(type)
+      const filter = type === '' ? {} : (queryString.type = { type });
+      const filteredProductsPageData = await ProductService.getFilteredProducts(filter);
+      setPageState(filteredProductsPageData.data)
+      // return filteredProductsPageData.data;
     }
-  }
-
-  async getProductData(type) {
-    const filter = type === '' ? {} : (this.queryString.type = { type });
-    const filteredProductsPageData = await ProductService.getFilteredProducts(filter);
-
-    return filteredProductsPageData.data;
-  }
-
-  setPageState(filteredProductsPageData) {
-    if (filteredProductsPageData === undefined) {
-      return;
+  
+    const setPageState = (filteredProductsPageData) => {
+      if (filteredProductsPageData === undefined) {
+        return;
+      }
+      const typesList = filteredProductsPageData.types;
+      const brandsList = filteredProductsPageData.brands;
+      const productsList = filteredProductsPageData.products;
+      // this.setState({ productsList, typesList, brandsList, loading: false });
+      setTypesList(typesList);
+      setBrandsList(brandsList);
+      setProductsList(productsList);
+      setLoading(false);
     }
-    const typesList = filteredProductsPageData.types;
-    const brandsList = filteredProductsPageData.brands;
-    const productsList = filteredProductsPageData.products;
-    this.setState({ productsList, typesList, brandsList, loading: false });
-  }
-
-    onFilterChecked = async (e, value) => {
-        const isChecked = e.target.checked;
-        const dataType = e.target.getAttribute('id');
-        this.setQueryStringState(isChecked, dataType, value);
-
-        const apiCall = await ProductService.getFilteredProducts(this.queryString);
-        this.setState({ productsList: apiCall.data.products });
-  };
-
-  setQueryStringState(isChecked, dataType, value) {
-    if (isChecked) {
-      this.brand.push(dataType);
-      this.queryString.brand = this.brand;
-      this.queryString.type = this.queryString.type.type === undefined ?
-            this.queryString.type : this.queryString.type.type;
-
-    } else {
-      let index = this.queryString[value].indexOf(dataType);
-      if (index !== -1) {
-        this.queryString[value].splice(index, 1);
+  
+     const onFilterChecked = async (e, value) => {
+          const isChecked = e.target.checked;
+          const dataType = e.target.getAttribute('id');
+          setQueryStringState(isChecked, dataType, value);
+  
+          const apiCall = await ProductService.getFilteredProducts(queryString);
+          // setState({ productsList: apiCall.data.products });
+          setProductsList(apiCall.data.products)
+    };
+  
+    const setQueryStringState = (isChecked, dataType, value) => {
+      if (isChecked) {
+        brand.push(dataType);
+        queryString.brand = brand;
+        queryString.type = getType ? getType  : '';
+        queryString.type = queryString.type.type === undefined ?
+              queryString.type : queryString.type.type;
+      } else {
+        let index = queryString[value].indexOf(dataType);
+        if (index !== -1) {
+          queryString[value].splice(index, 1);
+        }
+        queryString.type = getType ? getType  : '';
       }
     }
-  }
-
-  render() {
-    const { loading, typesList, brandsList, productsList } = this.state;
     return (
       <Fragment>
         {loading ? (
           <LoadingSpinner />
         ) : (
           <List
-            onFilterChecked={this.onFilterChecked}
+            onFilterChecked={onFilterChecked}
             typesList={typesList}
             brandsList={brandsList}
             productsList={productsList}
@@ -98,6 +115,6 @@ class ListContainer extends Component {
       </Fragment>
     );
   }
-}
+// }
 
-export default withRouter(ListContainer);
+export default (ListContainer);
