@@ -1,4 +1,4 @@
-import { FORM_EMAIL, SAVE_USER, REMOVE_USER } from '../types/types';
+import { FORM_EMAIL, SAVE_USER, REMOVE_USER, THEME_CHANGE, GET_QUANTITY } from '../types/types';
 
 let userInfo = JSON.parse(localStorage.getItem('state'));
 
@@ -10,20 +10,27 @@ const initialDefaultState = {
             email: '',
             type: ''
         }
-    }
+    },
+    theme : false,
+    quantity : 0
 }
-
 const defaultState = userInfo ? { userInfo } : { ...initialDefaultState };
 
-export default (state = defaultState, action) => {
+const login = (state = defaultState, action) => {
     switch (action.type) {
         case FORM_EMAIL:
             return { ...state, ...action };
         case SAVE_USER:
             return { userInfo: action.userInfo };
+        case GET_QUANTITY:
+            return { ...state, quantity : action.quantity };
+        case THEME_CHANGE:
+            return { ...state, [action.field] : action.value };
         case REMOVE_USER:
             return { ...initialDefaultState };
         default:
             return defaultState;
     }
 };
+
+export default login;
