@@ -1,20 +1,22 @@
 import axios from "axios";
 
-require("dotenv").config();
+// require("dotenv").config();
 const settingsUrl = "/api/settings";
 
-// const APIUrl = process.env.REACT_APP_DEV_API_URL;
-// const APIUrlShoppingCart = process.env.REACT_APP_API_URL_SHOPPINGCART;
-// @TODO: Replace 'deploymentid' in the next two lines with whatever your ENVIRONMENT github secret value is
-const APIUrl = 'http://contoso-traders-productsdeploymentid.REGION.cloudapp.azure.com/v1';
-const APIUrlShoppingCart = 'https://contoso-traders-cartsdeploymentid.orangeflower-95b09b9d.REGION.azurecontainerapps.io/v1';
+// Note: The '{PRODUCTS_API_ENDPOINT}', '{CARTS_API_ENDPOINT}' tokens will be substituted by github workflow.
+const APIUrl = process.env.REACT_APP_APIURL;
+const APIUrlShoppingCart = process.env.REACT_APP_APIURLSHOPPINGCART;
 const UseB2C = process.env.REACT_APP_USEB2C;
 const B2cAuthority = process.env.REACT_APP_B2CAUTHORITY;
-const B2cClientId = process.env.REACT_APP_B2CCLIENTID;
-const B2cScopes = process.env.REACT_APP_B2CSCOPES;
+const B2cClientId =  process.env.REACT_APP_B2CCLIENTID;
+const B2cScopes = [process.env.REACT_APP_B2CSCOPES];
+const userEmail = localStorage.getItem('state') ? JSON.parse(localStorage.getItem('state')).userName : null;
 
 const _HeadersConfig = (token, devspaces = undefined) => {
   const headers = token ? { Authorization: `Bearer ${token}` } : {};
+  if(userEmail){
+    headers['x-tt-email'] = userEmail
+  }
   if (devspaces) {
     headers["azds-route-as"] = devspaces;
   }
